pyproject.toml
[project]
name = "file-search"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "colorama>=0.4.6",
    "mcp>=1.16.0",
    "meilisearch>=0.37.0",
    "openpyxl>=3.1.5",
    "pillow>=11.3.0",
    "pytesseract>=0.3.13",
    "python-docx>=1.2.0",
]
